% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.comparison.R
\name{barchart.comparison}
\alias{barchart.comparison}
\title{Creates a bar chart to compare two years}
\usage{
barchart.comparison(ORIGIN, DESTINATION, VARIABLE, CLASSIFICATION, YEAR, STEP)
}
\arguments{
\item{ORIGIN}{is the country code of origin (e.g. "chl" for Chile)}

\item{DESTINATION}{is the country code of origin (e.g. "chn" for China)}

\item{VARIABLE}{is the variable to visualize and it can be "imports", "exports" or "exchange" (trade exchange)}

\item{CLASSIFICATION}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{YEAR}{is the initial year and the OEC's API ranges from 1962 to 2014}

\item{STEP}{is the distance between the years to compare (e.g if my year is 2010 and the distance is 2, then the function creates a chart comparing 2006, 2008, 2010, 2012 and 2014)}
}
\value{
Creates an \code{HTML} file with a bar chart visualization that compares two given years.
}
\description{
Creates a bar chart to compare two years
}
\examples{
barchart.comparison(chl, chn, exports, 6, 2010, 2)
}
\keyword{functions}

