% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{cran_convert}
\alias{cran_convert}
\title{Convert Volumes To Units Of Uncleaned Herring}
\usage{
cran_convert(value, unit)
}
\arguments{
\item{value}{Numeric. The value of the measurement to convert from.}

\item{unit}{Character. The unit of the measurement to convert from. Must be
a recognised volume unit as per \code{units::valid_udunits()}.}
}
\value{
An object of class \code{units}.
}
\description{
Converts a supplied volume measurement to crans.
}
\examples{
\dontrun{
    # 170.5 L is 1 cran
    cran_convert(170.5, "L")
    }
}
