% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{cran_install}
\alias{cran_install}
\title{Install Cran Conversion}
\usage{
cran_install()
}
\value{
Nothing.
}
\description{
Create the cran measurement unit. Asks for user input. Runs on startup.
}
\seealso{
\code{\link{cran_remove}}
}
