% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{cran_remove}
\alias{cran_remove}
\title{Remove Cran Conversion}
\usage{
cran_remove()
}
\value{
Nothing.
}
\description{
Remove the cran measurement unit. Asks for user input.
}
\seealso{
\code{\link{cran_install}}
}
