% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfoldr-package.R
\docType{data}
\name{data15p}
\alias{data15p}
\title{Intersection ellipses parameters}
\format{A matrix of columns: \code{A} (major semi-axis length), \code{C} (minor semi-axis length),
	\code{S=C/A} (shape factor), \code{alpha} (polar angle in intersecting plane) and coordinates in the
 intersection plane \code{x,y}.}
\usage{
data(data15p)
}
\description{
The data set consists of section profile parameters (assumed to come from prolate spheroids) of ellipses
fitted to measured particles of an aluminium matrix composite [5] from metallographic analysis. The data
set can be used to reconstruct the trivariate spatial (prolate) spheroid distribution.
}
\author{
M. Baaske
}
\keyword{datasets}
