% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracfixr.r
\docType{package}
\name{FracFixR-package}
\alias{FracFixR-package}
\title{FracFixR: Fraction Correction Framework for RNA-seq Data}
\description{
A compositional statistical framework for absolute proportion estimation 
between fractions in RNA sequencing data. FracFixR addresses the fundamental
challenge in fractionated RNA-seq experiments where library preparation and
sequencing depth obscure the original proportions of RNA fractions.
}
\author{
\strong{Maintainer}: Alice Cleynen \email{alice.cleynen@cnrs.fr} (\href{https://orcid.org/0000-0001-8083-0204}{ORCID})

Authors:
\itemize{
  \item Agin Ravindran
  \item Nikolay Shirokikh \email{nikolay.shirokikh@uwa.edu.au} (\href{https://orcid.org/0000-0001-8249-358X}{ORCID})
}

}
\keyword{internal}
