% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{getMaxRadius}
\alias{getMaxRadius}
\title{Maximum radius of exact simulated spheroids}
\usage{
getMaxRadius(S)
}
\arguments{
\item{S}{spheroid system, simulated by perfect simulation option}
}
\value{
maximum radius
}
\description{
Get maximum (random) radius
}
\details{
In case of exact simulation the maximum radius of all randomly generated radii is returned.
}
