% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdb.R
\name{getNMADB}
\alias{getNMADB}
\title{NMA catalog}
\usage{
getNMADB()
}
\value{
A data.frame with the network meta-analyses included in [1]. 
Several characterstics related to the publications 
(e.g. Journal.Name, Title) are included in the data.frame.
}
\description{
Petropoulou et al. compiled a database of network meta-analyses to 
serve as a source for empirical studies [1].
The database is hosted in a REDcap database at the 
Institute of Social and Preventive Medicine (ISPM) in the University of Bern.
Function getNMADB downloads the list of networks included in [1].
}
\examples{
catalog = getNMADB()
#Networks that labeled Verified have outcome data 
#that allow the analysis to be repeated.
nmalist = catalog[catalog$Verified=="True",]
nmalist
}
\references{
[1] Petropoulou M. et al. 2016 Bibliographic study showed improving statistical methodology 
of network meta-analyses published between 1999 and 2015,
\emph{Journal of Clinical Epidemiology}, 20--28 
\cite{DOI:10.1016/j.jclinepi.2016.11.002}
}
