\name{hs92_2char}
\alias{hs92_2char}
\title{HS92 groups}
\docType{data}
\description{
HS92 groups. This file is used to create the visualizations.
}
\usage{hs92_2char}
\format{
  A data frame with 22 observations on the following 2 variables.
  \describe{
    \item{\code{group}}{Contains the H292 groups (e.g. animal products, vegetable products, etc)}
    \item{\code{group_id}}{Contains the associated codes of every group (e.g. animal products is 01)}
  }
}
\examples{
hs92_2char
}
\keyword{datasets}
