% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_d3plus.R
\name{install_d3plus}
\alias{install_d3plus}
\title{Installs D3 and D3plus}
\usage{
install_d3plus()
}
\value{
Copies a folder named \code{d3plus} to the working directory and it contains the \code{js} files and icons to make the visualizations
}
\description{
Installs D3 and D3plus
}
\examples{
install_d3plus()
}
\keyword{functions}

