% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qle-package.R
\docType{data}
\name{matclust}
\alias{matclust}
\title{Matern cluster process data}
\format{A list named `\code{matclust}` with entries `\code{qsd0}` (final QL model after estimation),
 `\code{OPT}` (results of estimation by function \code{qle}) and `\code{Stest}` (the Scoring test).}
\usage{
data(matclust)
}
\description{
An example data set of a fitted Matern cluster point process model for the `\code{redwood}`
data (see package \code{spatstat}).
}
\details{
The model was fitted by our quasi-likelihood estimation method.
}
\author{
M. Baaske
}
\keyword{datasets}
