% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{parameterEstimates}
\alias{parameterEstimates}
\title{Estimated spatial histogram data}
\usage{
parameterEstimates(H, breaks)
}
\arguments{
\item{H}{trivariate histogram}

\item{breaks}{breaks as obtained from \code{\link{setbreaks}}}
}
\value{
list of size, angle and shape parameters, see \code{\link{parameters3d}}
}
\description{
Get histogram data from estimated joint distribution
}
\details{
Given the estimated joint distribution in histogram form the function
returns a list of \code{breaks} replicates which equals the number of
estimated count data for each class.
}

