% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{planarSection}
\alias{planarSection}
\title{Sphere planar section}
\usage{
planarSection(S, d, intern = FALSE)
}
\arguments{
\item{S}{list of spheres, see \code{\link{simSphereSystem}}}

\item{d}{distance of the intersecting xy-plane to the origin}

\item{intern}{\code{intern=FALSE} (default) return all sections otherwise
only those which have their centers inside the intersection window}
}
\value{
vector of circle radii
}
\description{
Intersect sphere system
}
\details{
Given a sphere system obtained from \code{\link{simSphereSystem}}
the function returns the section radii from intersecting all spheres
stored in \code{S}.
}

