% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdb.R
\name{readByID}
\alias{readByID}
\title{Read NMA dataset}
\usage{
readByID(recid)
}
\arguments{
\item{recid}{Record id of network}
}
\value{
A list with the name (id), data (dataset), type (continuous, binary,
rate, survival), effect (type of measure: RR OR RR RD ...), 
format (long, wide, iv). 
\itemize{
\item 'long' refers to data where each row represents a study treatment arm
\item 'wide' refers to data where each row represents a study treatment comparison
\item 'iv' refers to an 'inverse variance' format, where a comparison specific 
  estimate of the treatment effect and its standard error are reported.  
}
}
\description{
ReadByID downloads the dataset and main characteristics 
of the specified network meta-analysis.
You can list all ids from the catalog by calling \code{getNMADB}.
}
