% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{setupSphereSystem}
\alias{setupSphereSystem}
\title{Setup sphere system}
\usage{
setupSphereSystem(S, pl = 0)
}
\arguments{
\item{S}{sphere system}

\item{pl}{print level}
}
\value{
\code{NULL}
}
\description{
Reinitialize sphere system after R workspace reloading
}
\details{
The internally stored sphere system has to be reinitialized
after R workspace reloading. Calling this function is needed only in case one desires to get profile
sections of the sphere system again which has been previously stored as an R (list) object and afterwards
reloaded.
}

