% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedareaplot.comparison.R
\name{stackedareaplot.comparison}
\alias{stackedareaplot.comparison}
\title{Creates a stacked area plot to compare two years}
\usage{
stackedareaplot.comparison(ORIGIN, DESTINATION, VARIABLE, CLASSIFICATION, YEAR)
}
\arguments{
\item{ORIGIN}{is the country code of origin (e.g. "chl" for Chile)}

\item{DESTINATION}{is the country code of origin (e.g. "chn" for China)}

\item{VARIABLE}{is the variable to visualize and it can be "imports", "exports" or "exchange" (trade exchange)}

\item{CLASSIFICATION}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{YEAR}{is the initial year and the OEC's API ranges from 1962 to 2014}
}
\value{
Creates an \code{HTML} file with a stacked area plot visualization that compares two given years.
}
\description{
Creates a stacked area plot to compare two years
}
\examples{
stackedareaplot.comparison(chl, chn, exports, 6, 2010)
}
\keyword{functions}

