% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subcellular_annotation}
\alias{subcellular_annotation}
\title{Subcellular fractionation annotation example}
\format{
A data frame with 12 rows and 4 columns:
\describe{
  \item{Sample}{Sample identifier}
  \item{Condition}{Experimental condition (WT or Mutant)}
  \item{Type}{Fraction type (Total, Nuclear, or Cytoplasmic)}
  \item{Replicate}{Replicate identifier (Rep1 or Rep2)}
}
}
\source{
Simulated data generated for package examples
}
\usage{
subcellular_annotation
}
\description{
An annotation data frame for subcellular fractionation experiments
with nuclear and cytoplasmic fractions.
}
\examples{
data(subcellular_annotation)
head(subcellular_annotation)
}
\keyword{datasets}
